+++ ydb/core/protos/tx_datashard.proto
@@ -2404,7 +2404,7 @@ message TEvValidateUniqueIndexResponse {
 
     optional uint64 TabletId = 2;
 
-    optional Ydb.StatusIds.StatusCode Status = 3;
+    optional NKikimrIndexBuilder.EBuildStatus Status = 3;
     repeated Ydb.Issue.IssueMessage Issues = 4;
 
     optional uint64 RequestSeqNoGeneration = 5;
