+++ ydb/core/scheme/scheme_tablecell.h
@@ -426,6 +426,8 @@ inline ETriBool operator||(ETriBool a, ETriBool b) {
 }
 
 // Compare two TCell's for equality taking into account the NULL-value semantics.
+// If one of the cells is NULL, the result of comparison is NULL.
+// Else, the result is True when a == b and False when a != b.
 inline ETriBool TypedCellsEqualWithNullSemantics(const TCell& a, const TCell& b, const NScheme::TTypeInfoOrder& type) {
     if (a.IsNull() || b.IsNull()) {
         return ETriBool::Null;
@@ -435,6 +437,10 @@ inline ETriBool TypedCellsEqualWithNullSemantics(const TCell& a, const TCell& b,
 }
 
 // Compare two TCell vectors for equality taking into account the NULL-value semantics.
+// If there is no NULL fields from both sides, the result is True when all the fields are equal and False otherwise.
+// If all the fields that are NOT NULL are equal, but we have NULLs, the result is NULL.
+// If we have two NOT NULL fields that are not equal, the result is False.
+// See CompareWithNullSemantics test.
 template<class TTypeClass>
 inline ETriBool TypedCellVectorsEqualWithNullSemantics(const TCell* a, const TCell* b, const TTypeClass* type, const ui32 cnt) {
     ETriBool result = ETriBool::True;
